requirements.txt
accelerate==0.20.3
addict==2.4.0
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs 
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==22.2.0
backcall
beautifulsoup4==4.11.2
bleach==6.0.0
Brotli 
certifi
cffi==1.15.1
chamfer==2.0.0
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.3
ConfigArgParse==1.5.3
cycler==0.11.0
Cython==3.0.10
dash==2.9.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.6
decorator 
defusedxml==0.7.1
diffusers==0.21.4
distlib==0.3.6
docker-pycreds==0.4.0
easydict==1.13
einops==0.6.1
emd==0.0.0
emd-ext==0.0.0
entrypoints==0.4
exceptiongroup==1.2.1
fastjsonschema==2.16.3
filelock==3.10.0
Flask==2.2.3
flit_core
fonttools==4.38.0
freetype-py==2.3.0
frozenlist==1.3.3
fsspec==2023.1.0
gdown==4.6.4
geomloss==0.2.6
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.51.3
h5py==3.8.0
htmlgenerator==1.2.30
huggingface-hub==0.16.4
hydra-core==1.3.2
idna
imageio==2.26.0
importlib-metadata 
importlib-resources==5.12.0
iniconfig==2.0.0
ipykernel==6.16.2
ipython
ipython-genutils==0.2.0
ipywidgets==8.0.4
itsdangerous==2.1.2
jedi
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.17.3
jupyter-server==1.24.0
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
KNN-CUDA
llvmlite==0.39.1
loguru==0.6.0
Mako
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline 
mdurl==0.1.2
MiniSom==2.3.2
mistune==3.0.2
mkl-fft==1.3.1
mkl-random 
mkl-service==2.4.0
msgpack==1.0.5
munch 
nbclassic==1.1.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.6.3
ninja==1.11.1.1
notebook==6.5.7
notebook_shim==0.2.4
numba 
numpy 
omegaconf==2.3.0
open3d==0.9.0.0
opencv-python==4.7.0.72
opencv-python-headless==4.10.0.84
packaging==23.0
pandas==1.3.5
pandocfilters==1.5.1
parso 
pexpect 
pickleshare 
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.1.1
plotly==5.13.1
pluggy==1.2.0
plyfile==0.9
pointnet2==0.0.0
pointnet2-ops 
pooch==1.8.2
prometheus-client==0.17.1
prompt-toolkit 
protobuf==3.20.3
psutil==5.9.4
ptyprocess 
pycparser==2.21
pycuda k
pyembree==0.1.6
pyglet==1.5.16
Pygments==2.17.2
pymesh2 
pymeshlab==2022.2.post4
PyOpenGL==3.1.0
pyparsing==3.0.9
pyquaternion==0.9.9
pyrender==0.1.45
pyrsistent==0.19.3
PySocks 
pysom==0.1.4
pytest==7.4.4
python-dateutil==2.8.2
pytools
pytorch3d==0.7.0
pytz==2022.7.1
pyvista==0.38.6
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==25.0.1
ray==2.3.0
regex==2024.4.16
requests
rich==13.8.0
safetensors==0.4.4
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
scooby==0.7.3
Send2Trash==1.8.3
sentry-sdk==2.13.0
setproctitle==1.3.3
six 
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.4
tenacity==8.2.2
tensorboardX==2.6
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.4.5
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
torch==1.11.0+cu113
torch-cluster==1.6.1
torch-ema==0.3
torchaudio==0.11.0+cu113
torchsparse 
torchvision==0.12.0+cu113
tornado==6.2
tqdm 
traitlets 
transformers==4.30.2
transforms3d==0.4.1
trimesh==3.9.43
typing_extensions
urllib3
virtualenv==20.21.0
visdom==0.2.4
vtk==9.3.1
wandb==0.17.7
wcwidth 
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.2.3
widgetsnbextension==4.0.5
yacs==0.1.8
zipp 
